+++ jbrowse/components/common.tsx
