+++ jbrowse/plugins/ModifyMainMenu.tsx
