+++ jbrowse/utils/linear/dynamic.ts
